Moved functions from App.js into child components
