+++ src/components/MenuItem.js
@@ -4,8 +4,8 @@ import '../styles/MenuItem.scss';
 class MenuItem extends React.Component {
     constructor () {
         super();
-        this.handleSubmit = this.handleSubmit.bind(this);
         this.incQuantity = this.incQuantity.bind(this);
+        this.handleSubmit = this.handleSubmit.bind(this);
         this.state = {
             quantity: 0
         }
@@ -21,8 +21,25 @@ class MenuItem extends React.Component {
 
     handleSubmit (event) {
         event.preventDefault();
-        this.props.addToCurrentPurchase(this.state.quantity, this.props.item.id);
-    }
+        const id = this.props.item.id;
+        let currentPurchase;
+        const initialCurrentPurchase = { [id]: {id: id, quantity: this.state.quantity} }
+        if (!this.props.currentPurchase) {
+            currentPurchase = { items: initialCurrentPurchase }
+          } else {
+          if (this.state.quantity > 0) {
+            currentPurchase = Object.assign({}, this.props.currentPurchase, {
+              items: Object.assign({}, this.props.currentPurchase.items, initialCurrentPurchase )
+            });
+          } else {
+            currentPurchase = this.props.currentPurchase;
+            delete currentPurchase.items[id];
+          }
+        }
+        this.props.receiveCurrentPurchase(currentPurchase);
+        // name: "Dave",
+        // tel: "07901 972 811"
+      }
 
     render () {
 
@@ -33,11 +50,16 @@ class MenuItem extends React.Component {
             <img src={img}/>
             <ul>
                 <li>{this.props.item.item} &pound;{this.props.getCurrency(this.props.item.price)}</li>
-                <li><button onClick={(event) => this.incQuantity(-1, event)}> - </button> 
+                <li>
+                    <button onClick={(event) => this.incQuantity(-1, event)}> - </button> 
                      &nbsp;{this.state.quantity}&nbsp;
                     <button onClick={(event) => this.incQuantity(1, event)}> + </button> 
                 </li>
-                <li><button onClick={this.handleSubmit} type="submit">Add to order</button></li>
+                <li>
+                    <button onClick={this.handleSubmit} type="submit">
+                        Add to order
+                    </button>
+                </li>
             </ul>
         </section>)
     }
